Still adding more CC

+++ Commands/Utility/eventsping.js
@@ -1,21 +1,30 @@
-const discord = require('discord.js');
+const { RichEmbed } = require('discord.js');
 
 module.exports.run = async (bot,message,args) =>{
 
+    const settings = bot.sets;
 
-    let eventsRole = message.guild.roles.find('name', 'events');
+    if(args.length) return message.reply(new RichEmbed(bot.usages.get(exports.config.name)).setColor(settings.defaultEmbedColor))
+
+    let eventsRole = message.guild.roles.get(settings.eventsRole);
+
+    if(!eventsRole) return message.channel.send(new RichEmbed({
+        description: "Couldn't find the events role.",
+        color: settings.defaultEmbedColor
+    }))
     
     await eventsRole.edit({ mentionable : true });
 
-    message.channel.send(`<@&${eventsRole.id}>`);
+    await message.channel.send(`<@&${eventsRole.id}>`);
 
-    eventsRole.edit({ mentionable : false });
+    await eventsRole.edit({ mentionable : false });
 
-    message.delete();
-
-    return;
+    return message.delete();
 }
 
 module.exports.config = {
-    name: 'eventsping'
+    name: 'eventsping',
+    usage: "```.eventsping```",
+    desc: 'Mentions the events role.',
+    note: 'Make sure an events role has been set for the guild.'
 }
